Try to fix ImgWindow::SafeDelete() to not be so race/crashprone.
The original implementation was definitely harebrained - trying to delete textures and
the window in the middle of a render callback for that window was never a good idea.

Now we use a flight loop callback to tear down all pending windows.

That said, SafeDelete() is still a bad idea IMO - but at least now you can use it for
tearing down transients.